Weaken some language.
I've been thinking that, in order to set termination handlers in bin/www,
one would have to `require` the relevant modules into it that otherwise
wouldn't be needed in the file.

app.js, on the other hand, is the hub of many `require`s, having the most
references available in one file. That might be a better place to do
non-trivial termination handlers.